requirements.txt
libvirt-python == 5.3.0